Decompile func0f16d2ac

+++ src/include/types.h
@@ -6199,7 +6199,7 @@ struct var800aa5d8 {
 	s32 unk00;
 	s32 unk04;
 	s32 unk08;
-	s32 unk0c;
+	f32 unk0c;
 	s16 unk10;
 	s16 unk12;
 	s16 unk14;
